refactor(commands, refresher): 重构命令注册与数据刷新逻辑
- 移除未使用的函数和参数，简化命令注册过程
- 将数据刷新逻辑提取到新文件 `refresher.ts`，增强代码结构
- 引入定时器管理功能，优化状态栏显示逻辑
- 更新 `activate` 和 `deactivate` 方法以适应新的状态管理方式
